// src/types/generated/StreamerAction.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TransportToggleState } from "./streammagic_payloads/TransportToggleState";

export type StreamerAction = { "MuteSet": boolean } | "NextTrack" | "Pause" | "Play" | { "PlayPresetId": number } | { "PlayQueueId": number } | "PowerOn" | "PowerStandby" | "PowerToggle" | "PreviousTrack" | { "Seek": number } | { "SetRepeat": TransportToggleState } | { "SetShuffle": TransportToggleState } | { "SetSourceId": string } | "Stop" | "TogglePlayback" | { "VolumePercentSet": number } | { "VolumeStepChange": number } | { "VolumeStepSet": number };